modified controllers logic

+++ src/controllers/auth.ts
@@ -25,36 +25,34 @@ const login = asyncWrapper(
       return;
     }
     const { email, password } = req.body;
-    const user = User.findOne({ email });
+    const user = await User.findOne({ email });
     if (!user) {
       throw new BadRequestError("Wrong Credentials");
     }
+    const isPasswordCorrect = await bcrypt.compare(password, user.password);
+    if (!isPasswordCorrect) {
+      throw new BadRequestError("Invalid Credentials");
+    }
+    let phone = user.phone;
     const otp = generateOtp();
     await OTPHolder.deleteMany({ email });
 
     const otpHolder = new OTPHolder({
       email,
       otp,
+      password: user.password,
+      phone: user.phone,
+      name: user.name,
+      createdAt: new Date(),
     });
     await otpHolder.save();
-    const mailOptions = {
-      from: process.env.email,
-      to: email,
-      subject: "OTP for sign in",
-      text: `${otp} valid for 2 minutes`,
-    };
-    transport.sendMail(
-      mailOptions,
-      (err: Error | null, info: SentMessageInfo) => {
-        if (err) {
-          console.log(err);
-          next(err);
-          return;
-        } else {
-          res.status(200).json({ message: `OTP sent to ${email}` });
-        }
-      }
+    phone = phone.replace("+", "");
+    const message = await fetch(
+      `https://smsgw.tatatel.co.in:9095/campaignService/campaigns/qs?dr=false&sender=FRICOZ&recipient=${phone}&msg=Dear Customer, Your OTP for mobile number verification is ${otp}. Please do not share this OTP to anyone - Firstricoz Pvt. Ltd.&user=FIRSTR&pswd=First^01&PE_ID=1601832170235925649&Template_ID=1607100000000306120`
     );
+
+    const response = await message.json();
+    res.status(200).json({ message: `OTP sent sent to your phone ${phone} ` });
   }
 );
 
@@ -65,7 +63,7 @@ const verifyLoginOtp = asyncWrapper(
       res.status(400).json({ message: errors.array() });
       return;
     }
-    const { email, password, otp } = req.body;
+    const { email, otp } = req.body;
     const user = await User.findOne({ email });
     if (!user) {
       throw new BadRequestError("Invalid Credentails");
@@ -74,10 +72,6 @@ const verifyLoginOtp = asyncWrapper(
     const otpHolder = await OTPHolder.findOne({ email });
     if (!otpHolder) {
       throw new BadRequestError("Invalid OTP");
-    }
-    const isPasswordCorrect = await bcrypt.compare(password, user.password);
-    if (!isPasswordCorrect) {
-      throw new BadRequestError("Invalid Credentials");
     }
     const token = user.generateToken();
     res.cookie("auth_token", token, {
